Improved reusability of header, created control-bar component
